ShowWise-home/requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Cors
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
